Split oml.rs into oml/{scanner,parser,writer}.rs
oml.rs had grown to 1185 lines with three cleanly separated sections
(scanner, parser, writer) already marked off by its own banner comments.
Move each section into its own submodule, matching the mod.rs-less
sibling-file convention already used by infer.rs/materialize.rs/
formats/xml.rs (a top-level file plus an oml/ directory for tests).

oml.rs stays the public entry point (`mod scanner; mod parser; mod
writer;` plus the four pub fns and the Codec impl), now under 150 lines.
Moved items become pub(super)/pub(crate) as needed; nothing private
becomes public and no crate::oml::* path changes. oml/tests.rs is
unchanged except for two `#[cfg(test)]`-gated re-imports (Scalar,
MAX_INT_DIGITS) added to oml.rs so `use super::*` keeps resolving them.

Pure file reorganization, verified against the pre-split content with a
normalized line diff (visibility keywords/comments/use lines stripped) --
no logic changed.

Closes #53
